polish(pu): reuse res_block of DI-engine

+++ core/rl_utils/mcts/game_buffer.py
@@ -481,7 +481,7 @@ class GameBuffer(Buffer):
             # td_steps = config.td_steps - delta_td
             # td_steps = np.clip(td_steps, 1, 5).astype(np.int)
             # TODO(pu):
-            td_steps = np.clip(config.td_steps, 1, max(1, traj_len - state_index)).astype(np.int)
+            td_steps = np.clip(config.td_steps, 1, max(1, traj_len - state_index)).astype(np.int32)
 
             # prepare the corresponding observations for bootstrapped values o_{t+k}
             # o[t+ td_steps, t + td_steps + stack frames + num_unroll_steps]
